Fix another flaky test
